discord_bot/requirements.txt
# Discord.py
discord.py>=2.5.0

# Async HTTP client (for game server communication)
aiohttp>=3.9.0
httpx>=0.27.0

# Flask-RESTX API
flask>=3.0.0
flask-restx>=1.3.0
flask-cors>=5.0.0
flask-sqlalchemy>=3.1.0

# PostgreSQL database
psycopg[binary]>=3.1.0

# JWT authentication
PyJWT>=2.8.0

# Environment variables
python-dotenv>=1.0.0

# Werkzeug for threaded server
werkzeug>=3.0.0

# Development/Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
